Added which-key, headline and toggleterm.
Added plugins allow user to be able to edit markdown files and then with keymaps to then call pandoc to convert markdown files to a file. (need work-around for google drive paths to work with toggleterm since in toggleterm users cannot escape the ", instead user can use python's gdown to download the files from google drive to usb and then make changes and reupload them to google drive, OR, not use the whichkey mappings for now until toggleterm makes an update).